requirements.txt
gspread==6.1.4
matplotlib==3.5.1
numpy==1.26.4
pandas==2.0.2
protobuf==3.20.3
streamlit==1.23.1
Pillow==9.5.0
